pyproject.toml
[tool.poetry]
name = "ductape"
version = "0.1.0"
description = ""
authors = ["mosguinz <mos.guinz@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
mosspy = "^1.0.9"
requests = "^2.32.3"
seedir = "^0.4.2"
typer = {extras = ["all"], version = "^0.12.3"}

[tool.poetry.dev-dependencies]
black = "^24.4.2"

[tool.black]
line-length = 100

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
ductape = "ductape.main:app"
zipcheck = "zipfile_check:main"

#
#[packages]
#mosspy = "*"
#requests = "*"
#
#[scripts]
#mos_moss = "python mos_moss.py"
#zipfile_check = "python zipfile_check"
#
#[dev-packages]
#black = "*"
